Small fixes.

+++ PGS4NET.Windows.Tests/CaptionTests.cs
@@ -12,7 +12,6 @@
 
 using System.Drawing;
 using PGS4NET.Captions;
-using Xunit;
 
 namespace PGS4NET.Windows.Tests;
 
@@ -25,8 +24,8 @@ public class CaptionTests
     {
         var limitedRange = false;
         var colorSpace = ColorSpace.Bt2020ColorSpace;
-        var timeStamp = new PgsTimeStamp((long)Rng.Next());
-        var duration = new PgsTimeStamp((long)Rng.Next());
+        var timeStamp = new PgsTimeStamp(Rng.Next());
+        var duration = new PgsTimeStamp(Rng.Next());
         int x = 17;
         int y = 23;
         int width = 1024;
